Update PreferenceUtil.kt : SharedPreference에 저장할 데이터 설정
